backend/requirements.txt
fastapi[standard]==0.115.0
supabase==2.9.1
PyJWT[crypto]==2.9.0
pydantic-settings==2.5.2
httpx==0.27.2
geopy==2.4.1
